Add wrap around
In the original version of this extension, when you tried to move right
on the final tab, it would create a tab. I prefer that the former action
cycles back to the beginning of the tab list, same with when you try to
move left on the first tab. This also has some small other changes.